+++ src/opts/containers.rs
@@ -567,6 +567,11 @@ impl ContainerCreateOptsBuilder {
         image => "image"
     );
 
+    impl_str_field!(
+        /// User-specified image OS
+        image_os => "image_os"
+    );
+
     impl_str_enum_field!(
         /// Indicates how image volumes will be created. The default if unset is
         /// [`anonymous`](ImageVolumeMode::Anonymous).